dom/media/mediasource/test/test_MultipleInitSegments_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: Append buffer with multiple init segments</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/mp4");
  const init = new Uint8Array(await fetchWithXHR("bipbop/bipbop_videoinit.mp4"));
  const segment1 = new Uint8Array(await fetchWithXHR("bipbop/bipbop_video1.m4s"));
  const segment2 = new Uint8Array(await fetchWithXHR("bipbop/bipbop_video2.m4s"));
  const data = [ init, segment1, init, segment2 ];
  const length = data.map(d => d.byteLength).reduce((a, b) => a + b, 0);
  const arrayBuffer = new Uint8Array(length);
  let pos = 0;
  for (const buffer of data) {
    arrayBuffer.set(buffer, pos);
    pos += buffer.byteLength;
  }
  await loadSegment(sb, arrayBuffer);
  // Since we are passing multiple segments in one buffer,
  // the first durationchange event from parsing the init
  // segment will be fired before updateend.
  const p = once(v, "durationchange");
  ms.endOfStream();
  await p;
  ok(v.duration, 1.601666);
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
